More work into the frontend, Add a method extension ToUserFriendlyString, fix a jsonproperty and add validation if a recipe that no instruction

+++ Models/Information/Temperature.cs
@@ -4,7 +4,7 @@ namespace RecipeMate.Models.Information;
 
 public class Temperature
 {
-    [JsonProperty("number")] public int number { get; set; }
+    [JsonProperty("number")] public decimal number { get; set; }
 
     [JsonProperty("unit")] public string unit { get; set; }
 }
